Moved io code out of server.js into a config file
